# MM-24972: make GetHubForUserId lock-less and zero-alloc