--- a/wsapi/api.go
+++ b/wsapi/api.go
@@ -22,6 +22,4 @@ func Init(s *app.Server) {
 	api.InitUser()
 	api.InitSystem()
 	api.InitStatus()
-
-	a.HubStart()
 }