@agnivade 2020-07-02
@lindy65 - This needs testing around all websocket related functionality on an HA setup.